load assets on spawn, periodically spawn formations
